feat (dagster)!: refactor of assets into seperate python files
