+++ data-science/dagster/tutorials/getting_started_etl_tutorial/etl_tutorial/partitions.py
@@ -1,0 +1,7 @@
+import dagster as dg
+
+monthly_partition = dg.MonthlyPartitionsDefinition(start_date="2024-01-01")
+
+product_category_partition = dg.StaticPartitionsDefinition(
+    ["Electronics", "Books", "Home and Garden", "Clothing"]
+)
