+++ data-science/dagster/tutorials/getting_started_etl_tutorial/etl_tutorial/schedules.py
@@ -1,0 +1,7 @@
+import dagster as dg
+
+weekly_update_schedule = dg.ScheduleDefinition(
+    name="analysis_update_job",
+    target=dg.AssetSelection.keys("joined_data").upstream(),
+    cron_schedule="0 0 * * 1",  # every Monday at midnight
+)
